backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
pydantic==2.5.0
python-multipart==0.0.6
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0
google-cloud-trace==1.12.0
psutil==5.9.6
prometheus-client==0.19.0